refactor: move replay test reporting under replay
