+++ src/replay/test/reporters/junit.ts
@@ -1,7 +1,7 @@
 import fs from 'node:fs';
 import path from 'node:path';
-import type { ReplaySuiteResult, ReplaySuiteTestResult } from '../daemon/types.ts';
-import { AppError } from '../kernel/errors.ts';
+import type { ReplaySuiteResult, ReplaySuiteTestResult } from '../../../daemon/types.ts';
+import { AppError } from '../../../kernel/errors.ts';
 import type { ReplayTestReporter } from './types.ts';
 import {
   appendOptionalLine,
