+++ src/replay/test/trace.ts
@@ -1,7 +1,7 @@
 import fs from 'node:fs';
 import path from 'node:path';
-import type { ReplaySuiteTestResult } from './daemon/types.ts';
-import { formatDurationSeconds } from './utils/duration-format.ts';
+import type { ReplaySuiteTestResult } from '../../daemon/types.ts';
+import { formatDurationSeconds } from '../../utils/duration-format.ts';
 
 type ReplayActionStartTrace = {
   type: 'replay_action_start';
